+++ UnityProject/Saneject/Assets/Plugins/Saneject/Runtime/Bindings/Asset/AssetBindingBuilder.cs
@@ -76,7 +76,7 @@ namespace Plugins.Saneject.Runtime.Bindings.Asset
 
         /// <summary>
         /// Qualifies this binding to apply only when the injection target is of the given type or a derived type.
-        /// The injection target is the <see cref="Component" /> that owns the field, property, or method
+        /// The injection target is the object that owns the field, property, or method
         /// marked with <see cref="Attributes.InjectAttribute" />.
         /// </summary>
         /// <typeparam name="TTarget">The target type this binding applies to.</typeparam>
@@ -89,7 +89,7 @@ namespace Plugins.Saneject.Runtime.Bindings.Asset
 
         /// <summary>
         /// Qualifies this binding to apply only when the injection target is one of the specified types or a derived type.
-        /// The injection target is the <see cref="Component" /> that owns the field, property, or method
+        /// The injection target is the object that owns the field, property, or method
         /// marked with <see cref="Attributes.InjectAttribute" />.
         /// </summary>
         /// <param name="targetTypes">One or more target <see cref="Type" /> objects to match against.</param>